Update manual_test_guides.md
Added a few updates from @Kadino's review and a link to a newly published graphics-audio test suite. The overall LKG process doc has yet to be put up for review, but once that is ready I will add a link to it in this doc for any who want to know more about the LKG process.

Signed-off-by: AMZN-Dk <83721425+AMZN-Dk@users.noreply.github.com>